차트 캘린더 스포츠 합치는 중
잘 안됨

+++ Front/app/src/main/java/com/example/capston_spotyup/Chart/Api/ChartApi.kt
@@ -7,6 +7,6 @@ interface ChartApi {
     @GET("/games/chart/{sportsId}")
     suspend fun getChartData(
         @Query("userId") userId: Long,
-        @Path("sportsId") sportsId: Long
+        @Path("sportsId") sportsId: Int
     ): Call<ChartResponse>
 }
